backend/requirements.txt
# Flask核心依赖
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6

# 数据库
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7  # PostgreSQL支持
pymysql==1.1.0  # MySQL支持

# 视频处理
opencv-python==4.8.1.78
imageio==2.31.5
Pillow==10.0.1

# 流媒体处理
ffmpeg-python==0.2.0  # FFmpeg Python包装器

# 深度学习框架
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0
onnx==1.14.1
onnxruntime==1.15.1

# 计算机视觉
ultralytics==8.0.196  # YOLO
detectron2==0.6  # Facebook的检测框架
transformers==4.33.2  # Hugging Face

# 数据处理
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# 网络请求
requests==2.31.0
aiohttp==3.8.5

# 工具库
python-dotenv==1.0.0
redis==4.6.0
celery==5.3.1
gunicorn==21.2.0

# 开发工具
pytest==7.4.2
black==23.7.0
flake8==6.0.0
